Introduce `IndexOpsMixin` and move basic functions.
