+++ databricks/koalas/missing/indexes.py
@@ -56,7 +56,6 @@ class _MissingPandasLikeIndex(object):
     argsort = unsupported_function('argsort')
     asof = unsupported_function('asof')
     asof_locs = unsupported_function('asof_locs')
-    astype = unsupported_function('astype')
     contains = unsupported_function('contains')
     copy = unsupported_function('copy')
     delete = unsupported_function('delete')
@@ -95,17 +94,12 @@ class _MissingPandasLikeIndex(object):
     is_numeric = unsupported_function('is_numeric')
     is_object = unsupported_function('is_object')
     is_type_compatible = unsupported_function('is_type_compatible')
-    isin = unsupported_function('isin')
-    isna = unsupported_function('isna')
-    isnull = unsupported_function('isnull')
     item = unsupported_function('item')
     join = unsupported_function('join')
     map = unsupported_function('map')
     max = unsupported_function('max')
     memory_usage = unsupported_function('memory_usage')
     min = unsupported_function('min')
-    notna = unsupported_function('notna')
-    notnull = unsupported_function('notnull')
     nunique = unsupported_function('nunique')
     putmask = unsupported_function('putmask')
     ravel = unsupported_function('ravel')
@@ -171,7 +165,6 @@ class _MissingPandasLikeMultiIndex(object):
     argsort = unsupported_function('argsort')
     asof = unsupported_function('asof')
     asof_locs = unsupported_function('asof_locs')
-    astype = unsupported_function('astype')
     contains = unsupported_function('contains')
     copy = unsupported_function('copy')
     delete = unsupported_function('delete')
@@ -214,17 +207,12 @@ class _MissingPandasLikeMultiIndex(object):
     is_numeric = unsupported_function('is_numeric')
     is_object = unsupported_function('is_object')
     is_type_compatible = unsupported_function('is_type_compatible')
-    isin = unsupported_function('isin')
-    isna = unsupported_function('isna')
-    isnull = unsupported_function('isnull')
     item = unsupported_function('item')
     join = unsupported_function('join')
     map = unsupported_function('map')
     max = unsupported_function('max')
     memory_usage = unsupported_function('memory_usage')
     min = unsupported_function('min')
-    notna = unsupported_function('notna')
-    notnull = unsupported_function('notnull')
     nunique = unsupported_function('nunique')
     putmask = unsupported_function('putmask')
     ravel = unsupported_function('ravel')
